Move supplier brief eligibility route
Move /suppliers/<supplier_id>/briefs/<brief_id> to
    /briefs/<brief_id>/services?supplier_id=<supplier_id>

The new route returns an array of services that are eligible for the
brief. This makes it more similar to the /services route. At the moment
it requires a supplier_id so it will either return 1 or 0 services. In
future it may be expanded return all eligible services for the brief.
